Create CODE
This Verilog code defines an asynchronous D flip-flop with additional asynchronous reset (rst) and set (set) inputs. When the rst signal is asserted (high), the flip-flop resets to 0. When the set signal is asserted (high), the flip-flop sets to 1. Otherwise, the output q follows the input d asynchronously. The flip-flop changes its state immediately upon the assertion of rst or set, regardless of the clock.